nrnbbcore_write reorg
